Move base2 to integer_bits.py. Convert argument to Integer before calling digits member function.
