+++ sage-code/boolean_cayley_graphs/integer_bits.py
@@ -18,6 +18,36 @@ AUTHORS:
 #*****************************************************************************
 
 
+from sage.rings.integer import Integer
+
+
+base2 = lambda dim, num: Integer(num).digits(2, padto=dim)
+r"""
+Map ``num`` to :math:`\mathbb{F}_2^{dim}` using lexicographical ordering.
+
+INPUT:
+
+- ``num`` -- non-negative integer. The value to be mapped.
+- ``dim`` -- positive integer. The Boolean dimension.
+
+OUTPUT:
+
+A list of 0,1 integer values of length ``dim``.
+
+EXAMPLES:
+
+::
+
+    sage: from boolean_cayley_graphs.integer_bits import base2
+    sage: base2(5,3)
+    [1, 1, 0, 0, 0]
+    sage: base2(3,5)
+    [1, 0, 1]
+    sage: base2(3,1)
+    [1, 0, 0]
+"""
+
+
 def parity(n):
     r"""
     Return the bit parity of a non-negative integer.
